Separamos las funcionalidades de firebase por archivos js en la carpeta firebase, funciona el evento con el console log en el evento de Registrar Cuenta en registro.js
