+++ src/Firebase/firebase-import.js
@@ -1,0 +1,2 @@
+import { getAuth, createUserWithEmailAndPassword, signInWithEmailAndPassword, onAuthStateChanged } from "https://www.gstatic.com/firebasejs/9.6.8/firebase-auth.js";
+export { getAuth, createUserWithEmailAndPassword, signInWithEmailAndPassword, onAuthStateChanged }
