+++ src/index.html
@@ -30,7 +30,8 @@
     </header>
     <div id="root"></div>
     <script type="module" src="Firebase/auth.js"></script>
-    <script type="module" src="Firebase/firestore.js"></script>
+    <script type="module" src="Firebase/firebase-controller.js"></script>
+    <script type="module" src="Firebase/firebase-import.js"></script>
     <script type="module" src="main.js"></script>
     
 </body>
